backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Embeddings & Vector Store
sentence-transformers==2.2.2
chromadb==0.4.18

# LLM Integration
google-generativeai==0.3.2
groq>=0.4.0

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# URL Parsing
validators==0.22.0
